==> R part/Step06_Ex_post_BART_MIA.Rmd <==
x---
title: "Ex-post BART-MIA"
author: "Francesca Micocci"
date: "6/29/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

We start by uploading the required packages and the result tables from the ML algorithms.

# Import 

Import the required libraries
```{r libraries,message=FALSE,warning=FALSE}
library(tidyverse)
library(ggplot2)
library(dplyr)
library(kableExtra)
library(caret)
library(PRROC)
library(writexl)
library(ggpattern)
library(ggpubr)
library(scales)
```

Import data
```{r import data}
setwd('/your/directory/')
load('Export_project/Data/results_no_na.RData')
load('Export_project/Data/results.RData')
load(Export_project/Data/exporting_history.RData')
```

# Check Misclassification

Compute the confusion matrix distinguishing by exporter type. Looking at the *False Positives* and *False Negatives* we have a sense of how good the algorithm is in correctly predicting the export status of firms with different exporting patterns. 

```{r confusionmatrix exp_type}
#Generate subsets of the original dataset results, based on export type
constant <-filter(results,export_type=="Constant")
never <-filter(results,export_type=="Non-exporter")
discontinuous <-filter(results,export_type=="Discontinuous")
start <-filter(results,export_type=="Switching exporters")
stop <-filter(results,export_type=="Switching non-exporters")
#For each export type generate the confusion matrix and rename rows and columns
constant<-table(factor(constant$export,levels=c("1","0")),factor(constant$bart_mia_pred,levels=c("1","0")))
colnames(constant)=c("predicted\nexporter","predicted\nnon-exporter")
rownames(constant)=c("actual\nexporter","actual\nnon-exporter")
never<-table(factor(never$export,levels=c("1","0")),factor(never$bart_mia_pred,levels=c("1","0")))
colnames(never)=c("predicted\nexporter","predicted\nnon-exporter")
rownames(never)=c("actual\nexporter","actual\nnon-exporter")
discontinuous<-table(factor(discontinuous$export,levels=c("1","0")),factor(discontinuous$bart_mia_pred,levels=c("1","0")))
colnames(discontinuous)=c("predicted\nexporter","predicted\nnon-exporter")
rownames(discontinuous)=c("actual\nexporter","actual\nnon-exporter")
start<-table(factor(start$export,levels=c("1","0")),factor(start$bart_mia_pred,levels=c("1","0")))
colnames(start)=c("predicted\nexporter","predicted\nnon-exporter")
rownames(start)=c("actual\nexporter","actual\nnon-exporter")
stop<-table(factor(stop$export,levels=c("1","0")),factor(stop$bart_mia_pred,levels=c("1","0")))
colnames(stop)=c("predicted\nexporter","predicted\nnon-exporter")
rownames(stop)=c("actual\nexporter","actual\nnon-exporter")
  
#Generate tables to be later aggregated 
constant.p <- ggtexttable(constant, 
                        theme=ttheme('classic',
             colnames.style = colnames_style(fill = "white",color = "black",face="bold",size=9),
             rownames.style = rownames_style(fill = "white",color = "black",face="bold",size=9)
             ))%>%
  tab_add_title(text = "Constant Exporters", face = "bold", padding = unit(2, "line")) 

never.p <- ggtexttable(never, 
                        theme=ttheme('classic',
             colnames.style = colnames_style(fill = "white",color = "black",face="bold",size=9),
             rownames.style = rownames_style(fill = "white",color = "black",face="bold",size=9)
             ))%>%
  tab_add_title(text = "Non-Exporters", face = "bold", padding = unit(2, "line")) 

start.p <- ggtexttable(start, 
                        theme=ttheme('classic',
             colnames.style = colnames_style(fill = "white",color = "black",face="bold",size=9),
             rownames.style = rownames_style(fill = "white",color = "black",face="bold",size=9)
             ))%>%
  tab_add_title(text = "Switching exporters", face = "bold", padding = unit(2, "line")) 
stop.p <- ggtexttable(stop, 
                        theme=ttheme('classic',
             colnames.style = colnames_style(fill = "white",color = "black",face="bold",size=9),
             rownames.style = rownames_style(fill = "white",color = "black",face="bold",size=9)
             ))%>%
  tab_add_title(text = "Switching non-exporters", face = "bold", padding = unit(2, "line")) 

disc.p <- ggtexttable(discontinuous, 
                        theme=ttheme('classic',
             colnames.style = colnames_style(fill = "white",color = "black",face="bold",size=9),
             rownames.style = rownames_style(fill = "white",color = "black",face="bold",size=9)
             ))%>%
  tab_add_title(text = "Discontinuous exporters", face = "bold", padding = unit(2, "line")) 
# Generate the summary table 
Tables<-ggarrange(constant.p,never.p,start.p,stop.p,disc.p,
          ncol = 2, nrow = 3,
                    widths =   c(5,5))
#Remove useless stuff
rm(constant.p,disc.p,never.p,start.p,stop.p,constant,discontinuous,never,start,stop)
```
Display the Summary Table
```{r summary table}
Tables
```

Generate probability classes
```{r prob class}
results$prob_class<-ifelse(results$bart_mia_prob<0.10,"0-0.9",ifelse(results$bart_mia_prob>=0.10&results$bart_mia_prob<0.20,"0.1-0.19",ifelse(results$bart_mia_prob>=0.20&results$bart_mia_prob<0.30,"0.2-0.29",ifelse(results$bart_mia_prob>=0.30&results$bart_mia_prob<0.40,"0.3-0.39",ifelse(results$bart_mia_prob>=0.40&results$bart_mia_prob<0.50,"0.4-0.49",ifelse(results$bart_mia_prob>=0.50&results$bart_mia_prob<0.60,"0.5-0.59",ifelse(results$bart_mia_prob>=0.60&results$bart_mia_prob<0.70,"0.6-0.69",ifelse(results$bart_mia_prob>=0.70&results$bart_mia_prob<0.80,"0.7-0.79",ifelse(results$bart_mia_prob>=0.80&results$bart_mia_prob<0.90,"0.8-0.89","0.9-1")))))))))
```

Check confusion matrix by probability class, i.e. check the number of **actual exporters** and **actual non-exporters** according to the probability class (the probabilities are the output of the BART-MIA prediction) such observation belong to.
```{r exporers by prob class}
table(results$prob_class,results$export)%>%
  knitr::kable(digits = 4,
             col.names = c("number of non-exporters",
                           "numbe of exporters"),
             format.args = list(big.mark = ",")) %>%
  kableExtra::kable_classic(full_width = F,html_font = "Cambria")
```

Now we check how predictions vary with different exporter types. In particular, for each exporter type, we show the distribution of firms in probability class and we identify thassifieed observations when we use $0.5$ as probability threshold for the prediction of the **export status**.

```{r misclassified}
results$classification<-ifelse(results$bart_mia_pred==results$export,NA,"misclassified")

p <- results %>%
  ggplot( aes(x=prob_class, fill=as.factor(export),color=classification)) +
    geom_bar(linetype="dotted") +facet_wrap(~export_type)+geom_vline(xintercept = 5.5,
                color = "black", size=1,linetype="dashed")+ scale_colour_manual(values='black',na.translate=FALSE)+scale_fill_manual(values = c("grey87", "grey68")) +guides(color = guide_legend(override.aes = list(fill = "white"))) +
    theme_bw()+labs(x="Probability class")+scale_x_discrete(guide = guide_axis(angle = 45))
  
p
```


# Spearman correlation
Compute the Spearman correlation matrix of the predictions of the algorithms
```{r spearman corr}
predictors<-merge(results_no_na[,c("bvdidnumber","year","bart.prob","rf.prob","logit.prob","lasso.prob","export")],results[,c("bvdidnumber","year","bart_mia_prob")],on=c("bvdidnumber","year"))

Spearman_corr<-cor(predictors[,c("bart.prob","rf.prob","logit.prob","lasso.prob","bart_mia_prob")], method = "spearman")

Spearman_corr<-round(Spearman_corr,4)
```

Display the correlation matrix
```{r spearman table}
Spearman_corr[lower.tri(Spearman_corr)]<-""

Spearman_corr%>%
  knitr::kable(digits = 4)%>%
  kableExtra::kable_classic(full_width = F,html_font = "Cambria")
```
# Heterogeneous Exporting patterns

Check the performance of the algorithm for different exporting pattern.
```{r firm_type CM}
#Subset the original data by exporer type
always<-subset(results,export_type=="Constant")
never<-subset(results,export_type=="Non-exporter")
started<-subset(results,export_type=="Switching exporters")
stopped<-subset(results,export_type=="Switching non-exporters")
discontinuous<-subset(results,export_type=="Discontinuous")

#Manually compute Sensitivity for constant exporters
always.sensitivity<-table(always$bart_mia_pred,always$export)
always.sensitivity<-always.sensitivity[2,1]/(always.sensitivity[2,1]+always.sensitivity[1,1])
#Manually compute Specificity for non-exporters
never.specificity<-table(never$bart_mia_pred,never$export)
never.specificity<-never.specificity[1,1]/(never.specificity[1,1]+never.specificity[2,1])
#Generate confusion matrix for the remaining categories
cm_started=confusionMatrix(data=as.factor(started$bart_mia_pred),reference = as.factor(started$export),positive = "1")

cm_stopped=confusionMatrix(data=as.factor(stopped$bart_mia_pred),reference = as.factor(stopped$export),positive = "1")

cm_discontinuous=confusionMatrix(data=as.factor(discontinuous$bart_mia_pred),reference = as.factor(discontinuous$export),positive = "1")
# Generate vectors to be used for the final table.They contain the measures of performance
always.s<-c("constant",always.sensitivity,NA,NA,NA,NA,nrow(always))
never.s<-c("non-exporter",NA,never.specificity,NA,NA,NA,nrow(never))
started.s<-c("started",cm_started[["byClass"]][["Sensitivity"]],cm_started[["byClass"]][["Specificity"]],cm_started[["byClass"]][["Balanced Accuracy"]],NA,NA,nrow(started))
stopped.s<-c("stopped",cm_stopped[["byClass"]][["Sensitivity"]],cm_stopped[["byClass"]][["Specificity"]],cm_stopped[["byClass"]][["Balanced Accuracy"]],NA,NA,nrow(stopped))
discontinuous.s<-c("irregular",cm_discontinuous[["byClass"]][["Sensitivity"]],cm_discontinuous[["byClass"]][["Specificity"]],cm_discontinuous[["byClass"]][["Balanced Accuracy"]],NA,NA,nrow(discontinuous))
#Generate the table of the statistics
Performance_export_shape<-data.frame(matrix(vector(), 1, 7,
                dimnames=list(c(), c("Export_shape", "Sensitivity", "Specificity","Balanced_Accuracy","ROC","PR","Num. Obs"))),
                stringsAsFactors=F)
#Fill the table with the info above
Performance_export_shape<-rbind(Performance_export_shape,always.s,never.s,started.s,stopped.s,discontinuous.s)
#remove useless stuff
rm(cm_discontinuous,cm_stopped,cm_started,started.s,stopped.s,discontinuous.s,always.s,never.s,
  always.sensitivity,never.specificity)

```
Add ROC and PR
```{r ROC PR exp_type}
#Compute ROC and PR and insert them in the Summary table
roc.started<-roc.curve(scores.class0=as.numeric(started$bart_mia_prob),weights.class0 = as.numeric(started$export),
                     curve = T)
Performance_export_shape[4,5]=roc.started[["auc"]]
pr.started<-pr.curve(scores.class0=as.numeric(started$bart_mia_prob),weights.class0 = as.numeric(started$export),
                     curve = T)
Performance_export_shape[4,6]=pr.started[["auc.integral"]]

roc.stopped<-roc.curve(scores.class0=as.numeric(stopped$bart_mia_prob),weights.class0 = as.numeric(stopped$export),
                     curve = T)
Performance_export_shape[5,5]=roc.stopped[["auc"]]
pr.stopped<-pr.curve(scores.class0=as.numeric(stopped$bart_mia_prob),weights.class0=as.numeric(stopped$export),curve = T)
Performance_export_shape[5,6]=pr.stopped[["auc.integral"]]

roc.disc<-roc.curve(scores.class0=as.numeric(discontinuous$bart_mia_prob),weights.class0 = as.numeric(discontinuous$export),
                     curve = T)
Performance_export_shape[6,5]=roc.disc[["auc"]]
pr.disc<-pr.curve(scores.class0=as.numeric(discontinuous$bart_mia_prob),weights.class0=as.numeric(discontinuous$export),curve = T)
Performance_export_shape[6,6]=pr.disc[["auc.integral"]]
#Remove the first row, which is empty
Performance_export_shape<-Performance_export_shape[-1,]

#Remove useless stuff
rm(roc.disc,roc.started,roc.stopped,pr.disc,pr.started,pr.stopped)
```

Show the table
```{r summary exp_type}
#Transform the statistics in integers, that we can later round
for (i in 2:7){
  Performance_export_shape[,i]<-as.numeric(Performance_export_shape[,i])
}
#Show the final table
Performance_export_shape %>% 
  mutate_if(is.numeric, ~round(., 4))%>%
  knitr::kable(digits = 4,format.args = list(big.mark = ","),row.names = FALSE)%>%
  kableExtra::kable_classic(full_width = F,html_font = "Cambria")
```


## Switchin Exporters by year
```{r switching exporters}
#Generate a variable accounting for the year in which the firm started exporting
started$year.s<-ifelse(started$num_years==1,"2018",
                ifelse(started$num_years==2,"2017",
                ifelse(started$num_years==3,"2016",
                ifelse(started$num_years==4,"2015",
                ifelse(started$num_years==5,"2014",
                ifelse(started$num_years==6,"2013",
                ifelse(started$num_years==7,"2012",
                ifelse(started$num_years==8,"2011",NA
                       ))))))))
#Compute the confusion matrices of switchin exporters by year of switch
cm_started_2011=confusionMatrix(data=as.factor(started[started$year.s==2011,]$bart_mia_pred),reference = as.factor(started[started$year.s==2011,]$export),positive = "1")
cm_started_2012=confusionMatrix(data=as.factor(started[started$year.s==2012,]$bart_mia_pred),reference = as.factor(started[started$year.s==2012,]$export),positive = "1")
cm_started_2013=confusionMatrix(data=as.factor(started[started$year.s==2013,]$bart_mia_pred),reference = as.factor(started[started$year.s==2013,]$export),positive = "1")
cm_started_2014=confusionMatrix(data=as.factor(started[started$year.s==2014,]$bart_mia_pred),reference = as.factor(started[started$year.s==2014,]$export),positive = "1")
cm_started_2015=confusionMatrix(data=as.factor(started[started$year.s==2015,]$bart_mia_pred),reference = as.factor(started[started$year.s==2015,]$export),positive = "1")
cm_started_2016=confusionMatrix(data=as.factor(started[started$year.s==2016,]$bart_mia_pred),reference = as.factor(started[started$year.s==2016,]$export),positive = "1")
cm_started_2017=confusionMatrix(data=as.factor(started[started$year.s==2017,]$bart_mia_pred),reference = as.factor(started[started$year.s==2017,]$export),positive = "1")
cm_started_2018=confusionMatrix(data=as.factor(started[started$year.s==2018,]$bart_mia_pred),reference = as.factor(started[started$year.s==2018,]$export),positive = "1")
#Compute ROC and PR
roc.started.2011<-roc.curve(scores.class0=as.numeric(started[started$year.s==2011,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2011,]$export),
                     curve = T)
pr.started.2011<-pr.curve(scores.class0=as.numeric(started[started$year.s==2011,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2011,]$export),
                     curve = T)
roc.started.2012<-roc.curve(scores.class0=as.numeric(started[started$year.s==2012,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2012,]$export),
                     curve = T)
pr.started.2012<-pr.curve(scores.class0=as.numeric(started[started$year.s==2012,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2012,]$export),
                     curve = T)
roc.started.2013<-roc.curve(scores.class0=as.numeric(started[started$year.s==2013,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2013,]$export),
                     curve = T)
pr.started.2013<-pr.curve(scores.class0=as.numeric(started[started$year.s==2013,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2013,]$export),
                     curve = T)
roc.started.2014<-roc.curve(scores.class0=as.numeric(started[started$year.s==2014,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2014,]$export),
                     curve = T)
pr.started.2014<-pr.curve(scores.class0=as.numeric(started[started$year.s==2014,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2014,]$export),
                     curve = T)
roc.started.2015<-roc.curve(scores.class0=as.numeric(started[started$year.s==2015,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2015,]$export),
                     curve = T)
pr.started.2015<-pr.curve(scores.class0=as.numeric(started[started$year.s==2015,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2015,]$export),
                     curve = T)
roc.started.2016<-roc.curve(scores.class0=as.numeric(started[started$year.s==2016,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2016,]$export),
                     curve = T)
pr.started.2016<-pr.curve(scores.class0=as.numeric(started[started$year.s==2016,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2016,]$export),
                     curve = T)
roc.started.2017<-roc.curve(scores.class0=as.numeric(started[started$year.s==2017,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2017,]$export),
                     curve = T)
pr.started.2017<-pr.curve(scores.class0=as.numeric(started[started$year.s==2017,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2017,]$export),
                     curve = T)
roc.started.2018<-roc.curve(scores.class0=as.numeric(started[started$year.s==2018,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2018,]$export),
                     curve = T)
pr.started.2018<-pr.curve(scores.class0=as.numeric(started[started$year.s==2018,]$bart_mia_prob),weights.class0 = as.numeric(started[started$year.s==2018,]$export),
                     curve = T)
#Generate the rows to be added to the Summary Table
started.s.2011<-c("started 2011",cm_started_2011[["byClass"]][["Sensitivity"]],cm_started_2011[["byClass"]][["Specificity"]],cm_started_2011[["byClass"]][["Balanced Accuracy"]],roc.started.2011[["auc"]],pr.started.2011[["auc.integral"]],nrow(started[started$year.s==2011,]))
started.s.2012<-c("started 2012",cm_started_2012[["byClass"]][["Sensitivity"]],cm_started_2012[["byClass"]][["Specificity"]],cm_started_2012[["byClass"]][["Balanced Accuracy"]],roc.started.2012[["auc"]],pr.started.2012[["auc.integral"]],nrow(started[started$year.s==2012,]))
started.s.2013<-c("started 2013",cm_started_2013[["byClass"]][["Sensitivity"]],cm_started_2013[["byClass"]][["Specificity"]],cm_started_2013[["byClass"]][["Balanced Accuracy"]],roc.started.2013[["auc"]],pr.started.2013[["auc.integral"]],nrow(started[started$year.s==2013,]))
started.s.2014<-c("started 2014",cm_started_2014[["byClass"]][["Sensitivity"]],cm_started_2014[["byClass"]][["Specificity"]],cm_started_2014[["byClass"]][["Balanced Accuracy"]],roc.started.2014[["auc"]],pr.started.2014[["auc.integral"]],nrow(started[started$year.s==2014,]))
started.s.2015<-c("started 2015",cm_started_2015[["byClass"]][["Sensitivity"]],cm_started_2015[["byClass"]][["Specificity"]],cm_started_2015[["byClass"]][["Balanced Accuracy"]],roc.started.2015[["auc"]],pr.started.2015[["auc.integral"]],nrow(started[started$year.s==2015,]))
started.s.2016<-c("started 2016",cm_started_2016[["byClass"]][["Sensitivity"]],cm_started_2016[["byClass"]][["Specificity"]],cm_started_2016[["byClass"]][["Balanced Accuracy"]],roc.started.2016[["auc"]],pr.started.2016[["auc.integral"]],nrow(started[started$year.s==2016,]))
started.s.2017<-c("started 2017",cm_started_2017[["byClass"]][["Sensitivity"]],cm_started_2017[["byClass"]][["Specificity"]],cm_started_2017[["byClass"]][["Balanced Accuracy"]],roc.started.2017[["auc"]],pr.started.2017[["auc.integral"]],nrow(started[started$year.s==2017,]))
started.s.2018<-c("started 2018",cm_started_2018[["byClass"]][["Sensitivity"]],cm_started_2018[["byClass"]][["Specificity"]],cm_started_2018[["byClass"]][["Balanced Accuracy"]],roc.started.2018[["auc"]],pr.started.2018[["auc.integral"]],nrow(started[started$year.s==2018,]))
#Add the rows to the table
Performance_export_shape<-rbind.data.frame(Performance_export_shape,started.s.2011,started.s.2012,started.s.2013,started.s.2014,started.s.2015,started.s.2016,started.s.2017,started.s.2018)
#Remove uselessstuff
rm(cm_started_2011,cm_started_2012,cm_started_2013,cm_started_2014,cm_started_2015,cm_started_2016,cm_started_2017,cm_started_2018,roc.started.2011,roc.started.2012,roc.started.2013,roc.started.2014,roc.started.2015,roc.started.2016,roc.started.2017,roc.started.2018,pr.started.2011,pr.started.2012,pr.started.2013,pr.started.2014,pr.started.2015,pr.started.2016,pr.started.2017,pr.started.2018,started.s.2011,started.s.2012,started.s.2013,started.s.2014,started.s.2015,started.s.2016,started.s.2017,started.s.2018)
```
## Switching non-exporers by year
```{r switching non-exporters}
#Generate a variable accounting for the year in which the firm stopped exporting
stopped$year.s<-ifelse(stopped$num_years==1,"2011",
                ifelse(stopped$num_years==2,"2012",
                ifelse(stopped$num_years==3,"2013",
                ifelse(stopped$num_years==4,"2014",
                ifelse(stopped$num_years==5,"2015",
                ifelse(stopped$num_years==6,"2016",
                ifelse(stopped$num_years==7,"2017",
                ifelse(stopped$num_years==8,"2018",NA
                       ))))))))
#Compute confusion matrices for switching non-exporters, by year of switch
cm_stopped_2011=confusionMatrix(data=as.factor(stopped[stopped$year.s==2011,]$bart_mia_pred),reference = as.factor(stopped[stopped$year.s==2011,]$export),positive = "1")
cm_stopped_2012=confusionMatrix(data=as.factor(stopped[stopped$year.s==2012,]$bart_mia_pred),reference = as.factor(stopped[stopped$year.s==2012,]$export),positive = "1")
cm_stopped_2013=confusionMatrix(data=as.factor(stopped[stopped$year.s==2013,]$bart_mia_pred),reference = as.factor(stopped[stopped$year.s==2013,]$export),positive = "1")
cm_stopped_2014=confusionMatrix(data=as.factor(stopped[stopped$year.s==2014,]$bart_mia_pred),reference = as.factor(stopped[stopped$year.s==2014,]$export),positive = "1")
cm_stopped_2015=confusionMatrix(data=as.factor(stopped[stopped$year.s==2015,]$bart_mia_pred),reference = as.factor(stopped[stopped$year.s==2015,]$export),positive = "1")
cm_stopped_2016=confusionMatrix(data=as.factor(stopped[stopped$year.s==2016,]$bart_mia_pred),reference = as.factor(stopped[stopped$year.s==2016,]$export),positive = "1")
cm_stopped_2017=confusionMatrix(data=as.factor(stopped[stopped$year.s==2017,]$bart_mia_pred),reference = as.factor(stopped[stopped$year.s==2017,]$export),positive = "1")
cm_stopped_2018=confusionMatrix(data=as.factor(stopped[stopped$year.s==2018,]$bart_mia_pred),reference = as.factor(stopped[stopped$year.s==2018,]$export),positive = "1")
#Compute ROC and PR
roc.stopped.2011<-roc.curve(scores.class0=as.numeric(stopped[stopped$year.s==2011,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2011,]$export),
                     curve = T)
pr.stopped.2011<-pr.curve(scores.class0=as.numeric(stopped[stopped$year.s==2011,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2011,]$export),
                     curve = T)
roc.stopped.2012<-roc.curve(scores.class0=as.numeric(stopped[stopped$year.s==2012,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2012,]$export),
                     curve = T)
pr.stopped.2012<-pr.curve(scores.class0=as.numeric(stopped[stopped$year.s==2012,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2012,]$export),
                     curve = T)
roc.stopped.2013<-roc.curve(scores.class0=as.numeric(stopped[stopped$year.s==2013,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2013,]$export),
                     curve = T)
pr.stopped.2013<-pr.curve(scores.class0=as.numeric(stopped[stopped$year.s==2013,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2013,]$export),
                     curve = T)
roc.stopped.2014<-roc.curve(scores.class0=as.numeric(stopped[stopped$year.s==2014,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2014,]$export),
                     curve = T)
pr.stopped.2014<-pr.curve(scores.class0=as.numeric(stopped[stopped$year.s==2014,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2014,]$export),
                     curve = T)
roc.stopped.2015<-roc.curve(scores.class0=as.numeric(stopped[stopped$year.s==2015,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2015,]$export),
                     curve = T)
pr.stopped.2015<-pr.curve(scores.class0=as.numeric(stopped[stopped$year.s==2015,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2015,]$export),
                     curve = T)
roc.stopped.2016<-roc.curve(scores.class0=as.numeric(stopped[stopped$year.s==2016,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2016,]$export),
                     curve = T)
pr.stopped.2016<-pr.curve(scores.class0=as.numeric(stopped[stopped$year.s==2016,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2016,]$export),
                     curve = T)
roc.stopped.2017<-roc.curve(scores.class0=as.numeric(stopped[stopped$year.s==2017,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2017,]$export),
                     curve = T)
pr.stopped.2017<-pr.curve(scores.class0=as.numeric(stopped[stopped$year.s==2017,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2017,]$export),
                     curve = T)
roc.stopped.2018<-roc.curve(scores.class0=as.numeric(stopped[stopped$year.s==2018,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2018,]$export),
                     curve = T)
pr.stopped.2018<-pr.curve(scores.class0=as.numeric(stopped[stopped$year.s==2018,]$bart_mia_prob),weights.class0 = as.numeric(stopped[stopped$year.s==2018,]$export),
                     curve = T)
#Generate the rows to be added to the summary table
stopped.s.2011<-c("stopped 2011",cm_stopped_2011[["byClass"]][["Sensitivity"]],cm_stopped_2011[["byClass"]][["Specificity"]],cm_stopped_2011[["byClass"]][["Balanced Accuracy"]],roc.stopped.2011[["auc"]],pr.stopped.2011[["auc.integral"]],nrow(stopped[stopped$year.s==2011,]))
stopped.s.2012<-c("stopped 2012",cm_stopped_2012[["byClass"]][["Sensitivity"]],cm_stopped_2012[["byClass"]][["Specificity"]],cm_stopped_2012[["byClass"]][["Balanced Accuracy"]],roc.stopped.2012[["auc"]],pr.stopped.2012[["auc.integral"]],nrow(stopped[stopped$year.s==2012,]))
stopped.s.2013<-c("stopped 2013",cm_stopped_2013[["byClass"]][["Sensitivity"]],cm_stopped_2013[["byClass"]][["Specificity"]],cm_stopped_2013[["byClass"]][["Balanced Accuracy"]],roc.stopped.2013[["auc"]],pr.stopped.2013[["auc.integral"]],nrow(stopped[stopped$year.s==2013,]))
stopped.s.2014<-c("stopped 2014",cm_stopped_2014[["byClass"]][["Sensitivity"]],cm_stopped_2014[["byClass"]][["Specificity"]],cm_stopped_2014[["byClass"]][["Balanced Accuracy"]],roc.stopped.2014[["auc"]],pr.stopped.2014[["auc.integral"]],nrow(stopped[stopped$year.s==2014,]))
stopped.s.2015<-c("stopped 2015",cm_stopped_2015[["byClass"]][["Sensitivity"]],cm_stopped_2015[["byClass"]][["Specificity"]],cm_stopped_2015[["byClass"]][["Balanced Accuracy"]],roc.stopped.2015[["auc"]],pr.stopped.2015[["auc.integral"]],nrow(stopped[stopped$year.s==2015,]))
stopped.s.2016<-c("stopped 2016",cm_stopped_2016[["byClass"]][["Sensitivity"]],cm_stopped_2016[["byClass"]][["Specificity"]],cm_stopped_2016[["byClass"]][["Balanced Accuracy"]],roc.stopped.2016[["auc"]],pr.stopped.2016[["auc.integral"]],nrow(stopped[stopped$year.s==2016,]))
stopped.s.2017<-c("stopped 2017",cm_stopped_2017[["byClass"]][["Sensitivity"]],cm_stopped_2017[["byClass"]][["Specificity"]],cm_stopped_2017[["byClass"]][["Balanced Accuracy"]],roc.stopped.2017[["auc"]],pr.stopped.2017[["auc.integral"]],nrow(stopped[stopped$year.s==2017,]))
stopped.s.2018<-c("stopped 2018",cm_stopped_2018[["byClass"]][["Sensitivity"]],cm_stopped_2018[["byClass"]][["Specificity"]],cm_stopped_2018[["byClass"]][["Balanced Accuracy"]],roc.stopped.2018[["auc"]],pr.stopped.2018[["auc.integral"]],nrow(stopped[stopped$year.s==2018,]))
#Append the rows to the Summary table
Performance_export_shape<-rbind.data.frame(Performance_export_shape,stopped.s.2011,stopped.s.2012,stopped.s.2013,stopped.s.2014,stopped.s.2015,stopped.s.2016,stopped.s.2017,stopped.s.2018)
#Remove useless stuff
rm(cm_stopped_2011,cm_stopped_2012,cm_stopped_2013,cm_stopped_2014,cm_stopped_2015,cm_stopped_2016,cm_stopped_2017,cm_stopped_2018,roc.stopped.2011,roc.stopped.2012,roc.stopped.2013,roc.stopped.2014,roc.stopped.2015,roc.stopped.2016,roc.stopped.2017,roc.stopped.2018,pr.stopped.2011,pr.stopped.2012,pr.stopped.2013,pr.stopped.2014,pr.stopped.2015,pr.stopped.2016,pr.stopped.2017,pr.stopped.2018,stopped.s.2011,stopped.s.2012,stopped.s.2013,stopped.s.2014,stopped.s.2015,stopped.s.2016,stopped.s.2017,stopped.s.2018)
```
## Discontinuous exporters by number of exporting years
```{r discontinuous}
#Generate confusion matrices for Discontinuous exporters by number of exporting years
cm_discontinuous_1=confusionMatrix(data=as.factor(discontinuous[discontinuous$num_years==1,]$bart_mia_pred),reference = as.factor(discontinuous[discontinuous$num_years==1,]$export),positive = "1")
cm_discontinuous_2=confusionMatrix(data=as.factor(discontinuous[discontinuous$num_years==2,]$bart_mia_pred),reference = as.factor(discontinuous[discontinuous$num_years==2,]$export),positive = "1")
cm_discontinuous_3=confusionMatrix(data=as.factor(discontinuous[discontinuous$num_years==3,]$bart_mia_pred),reference = as.factor(discontinuous[discontinuous$num_years==3,]$export),positive = "1")
cm_discontinuous_4=confusionMatrix(data=as.factor(discontinuous[discontinuous$num_years==4,]$bart_mia_pred),reference = as.factor(discontinuous[discontinuous$num_years==4,]$export),positive = "1")
cm_discontinuous_5=confusionMatrix(data=as.factor(discontinuous[discontinuous$num_years==5,]$bart_mia_pred),reference = as.factor(discontinuous[discontinuous$num_years==5,]$export),positive = "1")
cm_discontinuous_6=confusionMatrix(data=as.factor(discontinuous[discontinuous$num_years==6,]$bart_mia_pred),reference = as.factor(discontinuous[discontinuous$num_years==6,]$export),positive = "1")
cm_discontinuous_7=confusionMatrix(data=as.factor(discontinuous[discontinuous$num_years==7,]$bart_mia_pred),reference = as.factor(discontinuous[discontinuous$num_years==7,]$export),positive = "1")
cm_discontinuous_8=confusionMatrix(data=as.factor(discontinuous[discontinuous$num_years==8,]$bart_mia_pred),reference = as.factor(discontinuous[discontinuous$num_years==8,]$export),positive = "1")
#Compute ROC and PR
roc.discontinuous.1<-roc.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==1,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==1,]$export),
                     curve = T)
pr.discontinuous.1<-pr.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==1,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==1,]$export),
                     curve = T)
roc.discontinuous.2<-roc.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==2,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==2,]$export),
                     curve = T)
pr.discontinuous.2<-pr.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==2,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==2,]$export),
                     curve = T)
roc.discontinuous.3<-roc.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==3,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==3,]$export),
                     curve = T)
pr.discontinuous.3<-pr.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==3,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==3,]$export),
                     curve = T)
roc.discontinuous.4<-roc.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==4,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==4,]$export),
                     curve = T)
pr.discontinuous.4<-pr.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==4,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==4,]$export),
                     curve = T)
roc.discontinuous.5<-roc.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==5,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==5,]$export),
                     curve = T)
pr.discontinuous.5<-pr.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==5,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==5,]$export),
                     curve = T)
roc.discontinuous.6<-roc.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==6,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==6,]$export),
                     curve = T)
pr.discontinuous.6<-pr.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==6,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==6,]$export),
                     curve = T)
roc.discontinuous.7<-roc.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==7,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==7,]$export),
                     curve = T)
pr.discontinuous.7<-pr.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==7,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==7,]$export),
                     curve = T)
roc.discontinuous.8<-roc.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==8,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==8,]$export),
                     curve = T)
pr.discontinuous.8<-pr.curve(scores.class0=as.numeric(discontinuous[discontinuous$num_years==8,]$bart_mia_prob),weights.class0 = as.numeric(discontinuous[discontinuous$num_years==8,]$export),
                     curve = T)
#Generate the rows
discontinuous.s.1<-c("1 year export",cm_discontinuous_1[["byClass"]][["Sensitivity"]],cm_discontinuous_1[["byClass"]][["Specificity"]],cm_discontinuous_1[["byClass"]][["Balanced Accuracy"]],roc.discontinuous.1[["auc"]],pr.discontinuous.1[["auc.integral"]],nrow(discontinuous[discontinuous$num_years==1,]))
discontinuous.s.2<-c("2 years export",cm_discontinuous_2[["byClass"]][["Sensitivity"]],cm_discontinuous_2[["byClass"]][["Specificity"]],cm_discontinuous_2[["byClass"]][["Balanced Accuracy"]],roc.discontinuous.2[["auc"]],pr.discontinuous.2[["auc.integral"]],nrow(discontinuous[discontinuous$num_years==2,]))
discontinuous.s.3<-c("3 years export",cm_discontinuous_3[["byClass"]][["Sensitivity"]],cm_discontinuous_3[["byClass"]][["Specificity"]],cm_discontinuous_3[["byClass"]][["Balanced Accuracy"]],roc.discontinuous.3[["auc"]],pr.discontinuous.3[["auc.integral"]],nrow(discontinuous[discontinuous$num_years==3,]))
discontinuous.s.4<-c("4 years export",cm_discontinuous_4[["byClass"]][["Sensitivity"]],cm_discontinuous_4[["byClass"]][["Specificity"]],cm_discontinuous_4[["byClass"]][["Balanced Accuracy"]],roc.discontinuous.4[["auc"]],pr.discontinuous.4[["auc.integral"]],nrow(discontinuous[discontinuous$num_years==4,]))
discontinuous.s.5<-c("5 years export",cm_discontinuous_5[["byClass"]][["Sensitivity"]],cm_discontinuous_5[["byClass"]][["Specificity"]],cm_discontinuous_5[["byClass"]][["Balanced Accuracy"]],roc.discontinuous.5[["auc"]],pr.discontinuous.5[["auc.integral"]],nrow(discontinuous[discontinuous$num_years==5,]))
discontinuous.s.6<-c("6 years export",cm_discontinuous_6[["byClass"]][["Sensitivity"]],cm_discontinuous_6[["byClass"]][["Specificity"]],cm_discontinuous_6[["byClass"]][["Balanced Accuracy"]],roc.discontinuous.6[["auc"]],pr.discontinuous.6[["auc.integral"]],nrow(discontinuous[discontinuous$num_years==6,]))
discontinuous.s.7<-c("7 years export",cm_discontinuous_7[["byClass"]][["Sensitivity"]],cm_discontinuous_7[["byClass"]][["Specificity"]],cm_discontinuous_7[["byClass"]][["Balanced Accuracy"]],roc.discontinuous.7[["auc"]],pr.discontinuous.7[["auc.integral"]],nrow(discontinuous[discontinuous$num_years==7,]))
discontinuous.s.8<-c("8 years export",cm_discontinuous_8[["byClass"]][["Sensitivity"]],cm_discontinuous_8[["byClass"]][["Specificity"]],cm_discontinuous_8[["byClass"]][["Balanced Accuracy"]],roc.discontinuous.8[["auc"]],pr.discontinuous.8[["auc.integral"]],nrow(discontinuous[discontinuous$num_years==8,]))
#Append the rows
Performance_export_shape<-rbind.data.frame(Performance_export_shape,discontinuous.s.1,discontinuous.s.2,discontinuous.s.3,discontinuous.s.4,discontinuous.s.5,discontinuous.s.6,discontinuous.s.7,discontinuous.s.8)
#Drop useless stuff
rm(cm_discontinuous_1,cm_discontinuous_2,cm_discontinuous_3,cm_discontinuous_4,cm_discontinuous_5,cm_discontinuous_6,cm_discontinuous_7,cm_discontinuous_8,roc.discontinuous.1,roc.discontinuous.2,roc.discontinuous.3,roc.discontinuous.4,roc.discontinuous.5,roc.discontinuous.6,roc.discontinuous.7,roc.discontinuous.8,pr.discontinuous.1,pr.discontinuous.2,pr.discontinuous.3,pr.discontinuous.4,pr.discontinuous.5,pr.discontinuous.6,pr.discontinuous.7,pr.discontinuous.8,discontinuous.s.1,discontinuous.s.2,discontinuous.s.3,discontinuous.s.4,discontinuous.s.5,discontinuous.s.6,discontinuous.s.7,discontinuous.s.8)
```
See the results:
```{r Final table}
#Trasnform the statistics in integers
for (i in 2:7){
  Performance_export_shape[,i]<-as.numeric(Performance_export_shape[,i])
}
#create a column to be used for sorting purposes
Performance_export_shape$sort<-c(1,2,3,12,21,4,5,6,7,8,9,10,11,13,14,15,16,17,18,19,20,22,23,24,25,26,27,28,29)
#Show the final table
Performance_export_shape %>% 
  mutate_if(is.numeric, ~round(., 4))%>%
  arrange(sort)%>%
  knitr::kable(digits = 4,row.names=FALSE,format.args = list(big.mark = ","))%>%
  kableExtra::remove_column(columns=8)%>%
  kableExtra::kable_classic(full_width = F,html_font = "Cambria")%>%
  kableExtra::row_spec(4:11, italic = TRUE,color="#666666")%>%
  kableExtra::row_spec(13:20, italic = TRUE,color="#666666")%>%
  kableExtra::row_spec(22:29, italic = TRUE,color="#666666")
```

## Békés and Murakozy (2012)

Follow the definition on Békés and Murakozy (2012), i.e. permanent spells corresponds to four consecutive exporting years or more.

1. Generate a factor variable taking values:
  + $2$ if the firm exported at 4 consecutive years 
  + $1$ if the firm exported less than 4 consecutive years
  + $0$ if the firm is a non-exporter
  
2. Add this factor variable to a subset of the original dataset
```{r BM def}
exporting_history$constant_shapes<-ifelse(apply(exporting_history[2:5], 1, sum)>=4|apply(exporting_history[3:6], 1, sum)>=4|apply(exporting_history[4:7], 1, sum)>=4|apply(exporting_history[5:8], 1, sum)>=4|apply(exporting_history[6:9], 1, sum)>=4|apply(exporting_history[7:10], 1, sum)>=4,2,ifelse(apply(exporting_history[2:10], 1, sum)==0,0,1))

BM_def<-merge(exporting_history[,c("bvdidnumber","export_shape","constant_shapes","export_type")],results[,c("bvdidnumber","year","export","bart_mia_pred","bart_mia_prob")],by="bvdidnumber")
```
Check the model performance for such firms
```{r BM types}
cm_permanent<-confusionMatrix(data = as.factor(BM_def[BM_def$constant_shapes=="2",]$bart_mia_pred),reference = as.factor(BM_def[BM_def$constant_shapes=="2",]$export),positive = "1")

roc_permanent<-roc.curve(scores.class0 = as.numeric(BM_def[BM_def$constant_shapes=="2",]$bart_mia_prob),weights.class0 =BM_def[BM_def$constant_shapes=="2",]$export,curve = TRUE)

pr_permanent<-pr.curve(scores.class0 = as.numeric(BM_def[BM_def$constant_shapes=="2",]$bart_mia_prob),weights.class0 =BM_def[BM_def$constant_shapes=="2",]$export,curve = TRUE)

cm_temporary<-confusionMatrix(data = as.factor(BM_def[BM_def$constant_shapes=="1",]$bart_mia_pred),reference = as.factor(BM_def[BM_def$constant_shapes=="1",]$export),positive = "1")

roc_temporary<-roc.curve(scores.class0 = as.numeric(BM_def[BM_def$constant_shapes=="1",]$bart_mia_prob),weights.class0 =BM_def[BM_def$constant_shapes=="1",]$export,curve = TRUE)

pr_temporary<-pr.curve(scores.class0 = as.numeric(BM_def[BM_def$constant_shapes=="1",]$bart_mia_prob),weights.class0 =BM_def[BM_def$constant_shapes=="1",]$export,curve = TRUE)

never.specificity<-table(BM_def[BM_def$constant_shapes=="0",]$bart_mia_pred,BM_def[BM_def$constant_shapes=="0",]$export)
never.specificity<-never.specificity[1,1]/(never.specificity[1,1]+never.specificity[2,1])

Perf_BM<-data.frame(matrix(vector(), 1, 7,
                dimnames=list(c(), c("Firm.Type", "Sensitivity", "Specificity","Balanced Accuracy","ROC","PR","Num_Obs"))),
                stringsAsFactors=F)

permanent<-c("Permanent",cm_permanent[["byClass"]][["Sensitivity"]],cm_permanent[["byClass"]][["Specificity"]],cm_permanent[["byClass"]][["Balanced Accuracy"]],roc_permanent[["auc"]],pr_permanent[["auc.integral"]],nrow(BM_def[BM_def$constant_shapes=="2",]))

temporary<-c("temporary",cm_temporary[["byClass"]][["Sensitivity"]],cm_temporary[["byClass"]][["Specificity"]],cm_temporary[["byClass"]][["Balanced Accuracy"]],roc_temporary[["auc"]],pr_temporary[["auc.integral"]],nrow(BM_def[BM_def$constant_shapes=="1",]))

never<-c("Non-exporters",NA,never.specificity,NA,NA,NA,nrow(BM_def[BM_def$constant_shapes=="0",]))

Perf_BM<-as.data.frame(rbind(Perf_BM,permanent,temporary,never))
Perf_BM<-subset(Perf_BM,!is.na(Perf_BM$Firm.Type))

rm(cm_permanent,cm_temporary,never.specificity,permanent,temporary,never,roc_permanent,roc_temporary,pr_permanent,pr_temporary)
```
Show the results
```{r}
#Trasnform the statistics in integers
for (i in 2:7){
  Perf_BM[,i]<-as.numeric(Perf_BM[,i])
}
Perf_BM%>%
  knitr::kable(digits = 4,row.names=FALSE,format.args = list(big.mark = ","))%>%
  kableExtra::kable_classic(full_width = F,html_font = "Cambria")

```


# Probability Distribution of Non-exporters

```{r prob distrib}
png("tiff_pic/Fig3_exporting_score.png", width = 1000, height = 600,units="px",res=150)
ggplot(results[which(results$export_type=="Non-exporter"),], aes(x=bart_mia_prob)) +
      geom_density(alpha=.22)+labs( x = "Predicted exporting score", 
      y = "density") +theme_bw()+geom_vline(xintercept=quantile(results[which(results$export_type=="Non-exporter"),]$bart_mia_prob,prob=0.5),linetype="dashed")
dev.off()
```

# Variable importance from BART_MIA

Download locally the variable importance dataset generated in Step5 and Load it here
```{r load var_imp}
setwd('/your/directory/')
load("Server/var_imp.RData")
```

Generate graphs
```{r var_imp custom}
var.imp<-as.data.frame(var.imp)
var.imp<-var.imp[var.imp$avg_var_props>0.0106,]
var.imp<-var.imp[order(-var.imp$avg_var_props),] 
var.imp$predictor<-c('external economies of scale','current assets','working capital','material costs','creditors','debtors','operating revenue turnover','financial expenses','taxation','intangible fixed assets ','age','nace 10','added value','patents: 0','size-age','shareholders funds','tangible fixed assets','industrial spillovers','nuts2: FR10','total assets','regional spillovers','financial revenue','nace 23','nace 32','fixed assets','EBITDA','current liabilities','inward FDI: 0', 'nuts2: FRFC2','outward FDI: 0','depreciation amortization','corporate control: 1','markups','nace 25')
var.imp$type<-c('internationalization','accounts','accounts','accounts','accounts','accounts','accounts','accounts','accounts', 'accounts','demographics','demographics', 'accounts','others','financial ratios','accounts','accounts','internationalization','demographics', 'accounts','internationalization','accounts','demographics','demographics', 'accounts','accounts','accounts','internationalization','demographics','internationalization','accounts','demographics','others','demographics')

```

Graph
```{r var_imp graph}
#Reorder by type of predictor and inclusion proportion
var.imp <-var.imp[order(var.imp$type,-var.imp$avg_var_props),]
var.imp$sort<-c(1:nrow(var.imp))
#remove the row names
rownames(var.imp) <- NULL
# identify Nace and Nuts
index.nace<-grep("nace",var.imp$predictor)
index.nuts<-grep("nuts",var.imp$predictor)
index.cat<-append(index.nace,index.nuts)
#remove Nace and Nuts
var.imp<-var.imp[-index.cat,]
#drop useless stuff
rm(index.cat,index.nace,index.nuts)
#Draw rthe graph
png("tiff_pic/Fig4_variable_inclusion_proportions.png", width = 1000, height = 600,units="px",res=120)
ggplot(var.imp, aes(x=reorder(predictor,sort), y=avg_var_props,fill=type)) + 
    geom_bar(position="dodge", stat="identity",colour="black")+
    geom_errorbar( aes(x=predictor, ymin=avg_var_props-sd_var_props, ymax=avg_var_props+sd_var_props), width=0.4, colour="#610000", alpha=0.9, size=1)+ theme_classic()+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),legend.position="bottom",legend.background = element_rect(size=0.5, linetype="solid", 
                                  colour ="black"))+ scale_fill_manual(values = c("gray100","gray75","gray50","gray88","gray25"))+ labs(fill = "Predictor Type")+
  xlab("Predictor") + ylab("Inclusion proportion")

dev.off()
```

# Score by Nace and Nuts

Generate the boxplot of the exporting score by Nace. 

1. Add to the dataset a variable containing the Labels of variable nace_2d

2. Generate the boxplot for the **non-exporters**

```{r boxlot non-exp by Nace}
results$Nace<-factor(results$nace_2d,
                        levels=c(10,11,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32),
                        labels = c("10: Food products","11: Beverages","13: Textiles","14: Wearing Apparel","15: Leather","16: Wood","17: Paper","18: Printing","19: Coke and refined petroleum","20: Chemicals","21: Pharmaceutical prod","22: Rubber and Plastic","23: Non-metallic mineral product","24: Basic metals","25: Fabricated metals","26: Computer, electronic and optical prod","27: Electrical equipment", "28: Machinery and equipment n.e.c.", "29: Motor vehicles","30: Other transportation",
"31: Furniture",
"32: Other manifacturing"))

results[results$export_type=="Non-exporter",] %>%
  mutate(name=fct_reorder(Nace,bart_mia_prob,.fun='median')) %>%
  ggplot(aes(x=reorder(factor(Nace),bart_mia_prob),y=bart_mia_prob))+geom_boxplot(outlier.shape=NA)+geom_hline(yintercept=quantile(results$bart_mia_prob,prob=0.5),size=1,color="gray70",linetype="longdash")+labs(x="Nace 2 digits",y="Exporting Score")+theme_bw()+coord_flip()

```

Export on excel the share of exporters with exporting score above 50th percentile, by nuts2. This is needed to use Tableau for maps.

1. Generate a table containing the number of exporter types by region

2. Generate a table containing the number of exporter with exporting score above the national median by region and types

3. Generate the final table, by merging the two above and keeping only non-exporters
```{r above_50}
firms_nuts1<-as.data.frame(table(results$nuts2,results$export_type))
firms_nuts2<-as.data.frame(table(results[results$bart_mia_prob>quantile(results$bart_mia_prob,prob=0.5),]$nuts2,results[results$bart_mia_prob>quantile(results$bart_mia_prob,prob=0.5),]$export_type))

firm_nuts<-merge(firms_nuts1,firms_nuts2,by=c("Var1","Var2"))
colnames(firm_nuts)<-c("Nuts1","export_type","total","above_50")
rm(firms_nuts1,firms_nuts2)

firm_nuts<-subset(firm_nuts,firm_nuts$export_type=="Non-exporter")
firm_nuts$loc_quot.num<-(firm_nuts$above_50/firm_nuts$total)
firm_nuts$tot_tot<-sum(firm_nuts$total)
firm_nuts$tot_above_50<-sum(firm_nuts$above_50)
firm_nuts$loc_quot.denom<-firm_nuts$tot_above_50/firm_nuts$tot_tot
firm_nuts$loc_quot<-firm_nuts$loc_quot.num/firm_nuts$loc_quot.denom
firm_nuts$share<-firm_nuts$above_50/firm_nuts$tot_above_50
firm_nuts$loc_quot.num<-NULL
firm_nuts$loc_quot.denom<-NULL
firm_nuts$tot_tot<-NULL
firm_nuts$tot_above_50<-NULL

write_xlsx(firm_nuts,'Export_project/50_pctile_share.xlsx')
```

Generate deciles of the distribution and plot the bar composition by industry
```{r industy_deciles}
get_decile <- function(x) ceiling(10*rank(x, ties.method="random") / length(x))

results$score_decile<-get_decile(results$bart_mia_prob)

score_by_nace<-as.data.frame(table(results$score_decile,results$Nace))

names(score_by_nace)<-c("Decile","Nace","Freq")
  
ggplot(data=score_by_nace,aes(x=Decile,y=Freq,fill=Nace,pattern=Nace))+ geom_bar_pattern(aes(pattern_angle=Nace),                                                                stat="identity",
       color='black',
       pattern_fill='black',
       pattern_spacing=0.02,
       pattern_density= 0.1,
       pattern_key_scale_factor=0.6)+scale_fill_grey(start = 0,
  end = 1) +theme_bw(9)+scale_pattern_manual(values =
          c('10: Food products'='none','11: Beverages'='stripe','13: Textiles'='circle','14: Wearing Apparel'='crosshatch','15: Leather'='none','16: Wood'='stripe','17: Paper'='circle','18: Printing'='crosshatch','19: Coke and refined petroleum'='none','20: Chemicals'='stripe','21: Pharmaceutical prod'='circle','22: Rubber and Plastic'='crosshatch','23: Non-metallic mineral product'='none','24: Basic metals'='stripe','25: Fabricated metals'='circle','26: Computer, electronic and optical prod'='crosshatch','27: Electrical equipment'='none','28: Machinery and equipment n.e.c.'='stripe','29: Motor vehicles'='circle','30: Other transportation'='crosshatch','31: Furniture'='none','32: Other manifacturing'='stripe'))+ scale_pattern_spacing_discrete(range = c(0.01, 0.05))+
  guides(fill=guide_legend(ncol=1))
```

# Check potential exporters

Generate the subset of potential exporters, i.e.firms which are Non-exporters and were misclassified. Select those in the last two probability classes.

```{r potential_exp}
potential_exporters<-subset(results, results$bart_mia_pred!=results$export&results$export_type=="Non-exporter")
potential_exporters<-subset(potential_exporters,potential_exporters$prob_class=="0.8-0.89"|potential_exporters$prob_class=="0.9-1")
```

Check the regional distribution and generate the table for Tableau containing n. of potential firms by region, n.of total firms by region, share of potential exporters out of total firms by region.

```{r pot_exp by nuts}
potential_exporters2<-as.data.frame(table(potential_exporters$nuts2))
names(potential_exporters2)[names(potential_exporters2) == "Freq"] <- "Potential exporter"

firms_by_nuts<-as.data.frame(table(results$nuts2))
names(firms_by_nuts)[names(firms_by_nuts) == "Freq"] <- "Total"
table_final<-merge(potential_exporters2,firms_by_nuts,on="nuts2")
table_final$`Share of potential exporters`<-(table_final$`Potential exporter`/table_final$Total)*100
table_final=table_final[table_final$`Potential exporter`>0,]

write_xlsx(table_final,"/Export_project/potential_exporters.xlsx")
rm(potential_exporters2,firms_by_nuts)
```

Now do the same thing for Nace.
```{r pot_exp by nace}
firms_by_nace.pe<-as.data.frame(table(potential_exporters$Nace))

firms_by_nace_tot<-as.data.frame(table(results$Nace))
names(firms_by_nace_tot)[names(firms_by_nace_tot) == "Freq"] <- "Total"
firms_by_nace.pe<-merge(firms_by_nace.pe,firms_by_nace_tot,on="Var1")
firms_by_nace.pe$Percentage<-(firms_by_nace.pe$Freq/firms_by_nace.pe$Total)*100
rm(firms_by_nace_tot)
```

Now produce the graphs of 
1) potential exporters by Nace
2) Share of potential exporters out of total firms by Nace

```{r pot_exp by nace num}
ggplot(firms_by_nace.pe, aes( y=Freq, x=Var1,na.rm = TRUE)) + 
  coord_flip()+geom_col(position="dodge") +
  labs(title = "Number of Potential exporters by Nace rev.2", y = "Count", x = "Nace Rev.2",fill="Export Dummy")+theme_bw()+theme(legend.position="bottom")+scale_y_continuous(labels = comma_format(big.mark = ".",
                                           decimal.mark = ","))+geom_text(aes(label = round(Freq, 2)),position = position_dodge(0.95),vjust = 0.5,hjust = -0.2,size=3)+ expand_limits(y = 21)
```
```{r pot_exp by nace share}
ggplot(firms_by_nace.pe, aes( y=Percentage, x=Var1,na.rm = TRUE)) + 
  coord_flip()+geom_col(position="dodge") +
  labs(title = "Share of Potential exporters over total firms (in % pts)", y = "Count", x = "Nace Rev.2",fill="Export Dummy")+theme_bw()+theme(legend.position="bottom")+scale_y_continuous(labels = comma_format(big.mark = ".",
                                           decimal.mark = ","))+geom_text(aes(label = round(Percentage, 2)),position = position_dodge(0.95),vjust = 0.5,hjust = -0.2,size=3)+ expand_limits(y = 1.2)
```

Draw some graphs

```{r graphs pred prob}
panel_fn <- function(x, y, ...)
{
    panel.xyplot(x, y, type="l",col="grey",lwd=1)
    panel.xyplot(x, y, type="smooth", col="black", lwd=2)
}

xyplot(bart_mia_prob ~ year, started[started$num_years==5,], panel=panel_fn,main="Predicted probabilities for firms exporting from 2016")
xyplot(bart_mia_prob ~ year, started[started$num_years==7,], panel=panel_fn,main="Predicted probabilities for firms exporting from 2014")

xyplot(bart_mia_prob ~ year, stopped[stopped$num_years==5,], panel=panel_fn,main="Predicted probabilities for firms exporting until 2016")
xyplot(bart_mia_prob ~ year, stopped[stopped$num_years==7,], panel=panel_fn,main="Predicted probabilities for firms exporting until 2014")
```
